check the value the method
